docs: update documentation and imports after components/themed rename
- Update CLAUDE.md with new themed components structure and Tamagui references
- Add new consolidated theme-guide.md documentation
- Update all component imports from @/components/base to @/components/themed
- Fix jest config transform ignore patterns for native-base compatibility
