+++ CLAUDE.md
@@ -35,22 +35,12 @@ This project uses **pnpm** as the package manager. For Expo-compatible packages,
 
 ### Theming System
 
-- All UI components should use themed wrappers (`ThemedView`, `ThemedText`)
-- Colors are defined in `src/constants/colors.ts` with light/dark variants
+- All pages should prioritize using components from `src/components/themed/` directory over React Native native components
+- Theme colors configured in `src/tamagui.config.ts` with Tamagui system
 - Use `useThemeColor` hook to access theme-aware colors
 - Components automatically adapt to system theme changes
 
-### Component Patterns
-
-```typescript
-// Example themed component structure
-import { useThemeColor } from '@/hooks/use-theme-color';
-
-export function MyComponent({ lightColor, darkColor, ...props }) {
-  const color = useThemeColor({ light: lightColor, dark: darkColor }, 'text');
-  // component implementation
-}
-```
+For detailed theming patterns and configuration, see **[docs/theme-guide.md](docs/theme-guide.md)**.
 
 ### Platform-specific Code
 
@@ -65,6 +55,7 @@ export function MyComponent({ lightColor, darkColor, ...props }) {
 - **Expo SDK**: 53.0.20
 - **TypeScript**: Strict mode enabled
 - **Navigation**: Expo Router v5 (file-based)
+- **UI Framework**: Tamagui for styled components and theming
 - **Animations**: react-native-reanimated v3
 - **Keyboard Management**: react-native-keyboard-controller for advanced keyboard handling
 
@@ -76,10 +67,15 @@ src/              # Source code directory
     (tabs)/       # Tab navigation screens
     _layout.tsx   # Root layout with theme provider
   components/     # Reusable components
-    ui/          # UI-specific components
-  constants/      # App constants (Colors, etc.)
-  hooks/         # Custom React hooks
-assets/          # Images, fonts, and static files (root level)
+    themed/       # Themed wrapper components (use these instead of RN components)
+    ui/           # UI-specific components
+  constants/      # App constants
+  hooks/          # Custom React hooks
+  tamagui.config.ts  # Tamagui theme configuration
+assets/           # Images, fonts, and static files (root level)
+docs/             # Documentation
+  theme-guide.md  # Comprehensive theming guide
+  testing-guide.md # Testing patterns and examples
 ```
 
 ## TypeScript Configuration
