+++ src/components/hello-wave.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated';
 
-import { ThemedText } from '@/components/base/themed-text';
+import { ThemedText } from '@/components/themed/themed-text';
 
 export function HelloWave() {
   const rotationAnimation = useSharedValue(0);
