+++ src/components/parallax-scroll-view.tsx
@@ -7,7 +7,7 @@ import Animated, {
   useScrollViewOffset,
 } from 'react-native-reanimated';
 
-import { ThemedView } from '@/components/base/themed-view';
+import { ThemedView } from '@/components/themed/themed-view';
 import { useBottomTabOverflow } from '@/components/ui/tab-bar-background';
 import { useColorScheme } from '@/hooks/use-color-scheme';
